+++ src/components/FhiAccordion/FhiAccordion.stories.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable import/no-extraneous-dependencies */
 /* eslint-disable react/jsx-props-no-spreading */
-import { faAngleDown } from '@fortawesome/free-solid-svg-icons';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
 import FhiAccordion from './FhiAccordion';
@@ -12,38 +11,46 @@ export default {
 
 const Template: ComponentStory<typeof FhiAccordion> = (args) => <FhiAccordion {...args} />;
 
-const DefaultHeader = () => <h4>Dette er en ustylet overskrift</h4>;
 const DefaultContent = () => <div><span>Dette er ustylet innhold</span></div>;
 
 const defaultArgs = {
   items: [
     {
       id: '1',
-      header: <DefaultHeader />,
+      header: 'Item #1',
       content: <DefaultContent />,
     },
     {
       id: '2',
-      header: <DefaultHeader />,
+      header: 'Item #2',
       content: <DefaultContent />,
     },
     {
       id: '3',
-      header: <DefaultHeader />,
+      header: 'Item #3',
       content: <DefaultContent />,
     },
     {
       id: '4',
-      header: <DefaultHeader />,
+      header: 'Item #4',
       content: <DefaultContent />,
     },
   ],
-  arrowIcon: faAngleDown,
 };
 
 export const Normal = Template.bind({});
-
 Normal.args = {
   items: defaultArgs.items,
-  arrowIcon: defaultArgs.arrowIcon,
+};
+
+export const Flush = Template.bind({});
+Flush.args = {
+  items: defaultArgs.items,
+  flush: true,
+};
+
+export const Alternate = Template.bind({});
+Alternate.args = {
+  items: defaultArgs.items,
+  alternate: true,
 };