+++ src/components/FhiHeader/FhiMainMenu.tsx
@@ -1,67 +1,62 @@
+/* eslint-disable jsx-a11y/click-events-have-key-events */
+/* eslint-disable jsx-a11y/no-static-element-interactions */
 import { FC, useState } from 'react';
 import { Link, NavLink } from 'react-router-dom';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 
 export type FhiMainMenuProps = {
   menuItems: {
     name: string,
     routerLink: string,
-  }[],
-  logo: string,
-  faIcons: {
-    menu: IconDefinition,
-    close: IconDefinition,
-  }
+  }[]
 };
 
-const FhiMainMenu:FC<FhiMainMenuProps> = ({ menuItems, logo, faIcons }) => {
+const FhiMainMenu:FC<FhiMainMenuProps> = ({ menuItems }) => {
   const [isOpen, setOpen] = useState(false);
 
   const mainMenuClose = () => setOpen(false);
   const mainMenuToggle = () => setOpen(!isOpen);
 
+  if (!menuItems || menuItems.length === 0) {
+    return null;
+  }
+
   return (
-    <div className="fhi-app__main-menu">
-      <div className="container fhi-app__main-menu-container">
-        {isOpen && <div className="fhi-backdrop-responsive" aria-hidden="true" onClick={mainMenuClose} />}
-        <nav className={`fhi-main-menu${isOpen ? ' fhi-main-menu--open' : ''}`}>
-          <div className="container fhi-main-menu__container">
-            <Link className="fhi-main-menu__home-link" to="/" onClick={mainMenuClose}>
-              <img
-                className="fhi-main-menu__home-link-img"
-                alt="FHI-logo"
-                src={logo}
-              />
-            </Link>
-            <button
-              className=" fhi-main-menu__toggler"
-              type="button"
-              aria-controls="fhi-main-menu"
-              aria-expanded={isOpen}
-              onClick={mainMenuToggle}
-            >
-              <FontAwesomeIcon icon={isOpen ? faIcons.close : faIcons.menu} />
-              <span className="fhi-main-menu__toggler-text">{ !isOpen ? 'Meny' : 'Lukk meny' }</span>
-            </button>
-            <div className={`collapse fhi-main-menu__collapse${isOpen ? ' show' : ''}`}>
-              <ul className="nav flex-column flex-xl-row fhi-main-menu__nav">
-                {menuItems.map((item) => (
-                  <li className="nav-item" key={item.name}>
-                    <NavLink
-                      className="nav-link fhi-main-menu__nav-link"
-                      to={item.routerLink}
-                      onClick={mainMenuToggle}
-                    >
-                      <span className="fhi-main-menu__nav-link-text">{ item.name }</span>
-                    </NavLink>
-                  </li>
-                ))}
-              </ul>
-            </div>
+    <div className="container fhi-header__main-menu-container">
+      <nav className={`fhi-main-menu ${isOpen ? 'fhi-main-menu--open' : ''}`}>
+        <div className="container fhi-main-menu__container">
+          <Link to="/" className="fhi-main-menu__home-link">
+            <i className="icon-fhi-logo fhi-header__logo-icon fhi-main-menu__logo-icon" />
+            <span className="visually-hidden">FHI Produktnavn</span>
+          </Link>
+          <button
+            type="button"
+            aria-controls="fhi-main-menu"
+            className="fhi-main-menu__toggler"
+            aria-expanded={isOpen}
+            onClick={mainMenuToggle}
+          >
+            <i className="icon-xmark fhi-main-menu__toggler-close-icon" />
+            <i className="icon-list fhi-main-menu__toggler-menu-icon" />
+            <span className="fhi-main-menu__toggler-text">{ !isOpen ? 'Meny' : 'Lukk' }</span>
+          </button>
+          <div className={`collapse fhi-main-menu__collapse ${isOpen ? 'show' : ''}`}>
+            <ul className="nav nav-tabs fhi-nav-tabs fhi-main-menu__nav">
+              {menuItems.map((link) => (
+                <li className="fhi-main-menu__nav-item" key={link.routerLink}>
+                  <NavLink
+                    to={link.routerLink}
+                    onClick={mainMenuClose}
+                    className="nav-link fhi-main-menu__nav-link"
+                  >
+                    {link.name}
+                  </NavLink>
+                </li>
+              ))}
+            </ul>
           </div>
-        </nav>
-      </div>
+        </div>
+      </nav>
+      {isOpen && <div className="fhi-backdrop-responsive" onClick={mainMenuClose} />}
     </div>
   );
 };